+++ assets/components/admin/TherapeutistsAdmin.vue
@@ -72,6 +72,44 @@
                 </div>
             </div>
 
+            <div class="form-group row">
+                <label for="confs" class="col-sm-2 col-form-label">Konferencje: </label>
+                <div class="col-sm-8">
+                    <input type="text" id="confs" class="form-control" v-model="act_conf">
+                </div>
+                <button type="button" class="col-sm-2 btn btn-success" @click="addConf(act_conf)">Dodaj</button>
+            </div>
+
+            <div class="form-group row">
+                <div class="col-sm-12">
+                    <table class="table table-hover">
+                        <tr><th>Dodane konferencje</th></tr>
+                        <tr v-for="(conf, index) in jsondata.conferences">
+                            <th>{{conf}}  <span class="deletebutton" @click="deleteConf(index)">x</span></th>
+                        </tr>
+                    </table>
+                </div>
+            </div>
+
+            <div class="form-group row">
+                <label for="publications" class="col-sm-2 col-form-label">Publikacje: </label>
+                <div class="col-sm-8">
+                    <input type="text" id="publications" class="form-control" v-model="act_pub">
+                </div>
+                <button type="button" class="col-sm-2 btn btn-success" @click="addPub(act_pub)">Dodaj</button>
+            </div>
+
+            <div class="form-group row">
+                <div class="col-sm-12">
+                    <table class="table table-hover">
+                        <tr><th>Dodane publikacje</th></tr>
+                        <tr v-for="(pub, index) in jsondata.publications">
+                            <th>{{pub}}  <span class="deletebutton" @click="deletePub(index)">x</span></th>
+                        </tr>
+                    </table>
+                </div>
+            </div>
+
 
             <div class="form-group row">
                 <label for="methods" class="col-sm-2 col-form-label">Metody: </label>
@@ -152,6 +190,8 @@
                 act_method: '',
                 act_school: '',
                 act_course: '',
+                act_pub: '',
+                act_conf: '',
                 therapeutists: [],
                 methods: [],
                 jsondata: {
@@ -162,7 +202,9 @@
                     education: [],
                     methods: [],
                     schools: [],
-                    courses: []
+                    courses: [],
+                    publications: [],
+                    conferences: []
                 },
                 selectedFile: null
             }
@@ -222,26 +264,34 @@
                     });
             },
             deleteTherapist(id) {
-                this.$http.delete(
-                    '/api/therapeutists/' + id,
-                    {name: this.jsondata.name},
-                    {
-                        headers: {}
-                    })
-                    .then(function (response) {
-                        console.log('Success!:', response.data);
-                        this.loading = false;
-                        this.fetchTherapeutists();
-                    }, function (response) {
-                        console.log('Error!:', response.data);
-                        this.loading = false;
-                    });
+                if (confirm("Czy na pewno chcesz usunąć terapeutę?")) {
+                    this.$http.delete(
+                        '/api/therapeutists/' + id,
+                        {name: this.jsondata.name},
+                        {
+                            headers: {}
+                        })
+                        .then(function (response) {
+                            console.log('Success!:', response.data);
+                            this.loading = false;
+                            this.fetchTherapeutists();
+                        }, function (response) {
+                            console.log('Error!:', response.data);
+                            this.loading = false;
+                        });
+                }
             },
             addMethod(method) {
                 this.jsondata.methods.push(method);
             },
             addSchool(school) {
                 this.jsondata.schools.push(school);
+            },
+            addPub(method) {
+                this.jsondata.publications.push(method);
+            },
+            addConf(school) {
+                this.jsondata.conferences.push(school);
             },
             addCourse(course) {
                 this.jsondata.courses.push(course);
@@ -251,6 +301,12 @@
             },
             deleteCourse(index) {
                 this.jsondata.courses.splice(index,1);
+            },
+            deleteConf(index) {
+                this.jsondata.conferences.splice(index,1);
+            },
+            deletePub(index) {
+                this.jsondata.publications.splice(index,1);
             },
             deleteMethod(index) {
                 this.jsondata.methods.splice(index,1);
